--- Second_Opinion_Express/config.js
@@ -1,3 +1,0 @@
-module.exports = {
-  secret: 'wHVfFHmcqAy49UmjnXEUdinoRzoUeESr',
-}
