+++ Second_Opinion_Express/db.js
@@ -4,8 +4,8 @@ const mysql = require('mysql2/promise')
 const pool = mysql.createPool({
   host: 'localhost',
   user: 'root',
-  password: 'manager',
-  database: 'so_db',
+  password: 'wwnbtr',
+  database: 'p2',
   waitForConnections: true,
   connectionLimit: 10,
   maxIdle: 10, // max idle connections, the default value is the same as `connectionLimit`
