+++ Second_Opinion_Express/package.json
@@ -1,11 +1,8 @@
 {
   "dependencies": {
     "cors": "^2.8.5",
-    "crypto-js": "^4.2.0",
     "express": "^4.18.2",
-    "jsonwebtoken": "^9.0.2",
     "morgan": "^1.10.0",
-    "multer": "^1.4.5-lts.1",
     "mysql2": "^3.9.1"
   }
 }
