+++ Second_Opinion_Java/src/main/resources/application.properties
@@ -4,7 +4,7 @@
 #server.servlet.context-path=/test_boot
 server.servlet.session.persistent=false
 #DB properties
-spring.datasource.url=jdbc:mysql://localhost:3306/p2?createDatabaseIfNotExist=true&useSSL=false&allowPublicKeyRetrieval=true
+spring.datasource.url=jdbc:mysql://localhost:3306/second_opinion_db?createDatabaseIfNotExist=true&useSSL=false&allowPublicKeyRetrieval=true
 spring.datasource.username=root
 spring.datasource.password=manager
 # JPA properties
